* commit '73034cf147f506eda48133c741488f27f538a307': (23 commits)
  Minor fix for gripper example.
  Add test for time steps given to control callbacks.
  Fix time step calculation.
  Duration: Remove move operator and ctor.
  Remove now-redundant test.
  Use Duration for GripperState.
  Duration: Add tests, docu.
  Add franka::Duration parameter to control callbacks.
  Fix tests.
  Clarify RobotState docu, run clang-format.
  Handle overflows for tick calculation.
  examples: Add comment regarding use of std::cout.
  test: Add assertion to avoid test hangs.
  test: Ignore realtime capabilities.
  Valgrind: Re-add origin tracking.
  Move tick handling into *Loop classes.
  MockServer: Check for shutdown in doForever.
  test: Turn of Valgrind origin tracking.
  MockServer: Reuse socket.
  MockServer: Add mutexes for TCP and UDP.
  ...

+++ src/robot_impl.h
@@ -63,7 +63,7 @@ class Robot::Impl : public RobotControl {
 
   research_interface::robot::MotionGeneratorMode motion_generator_mode_;
   research_interface::robot::ControllerMode controller_mode_;
-  uint32_t message_id_;
+  uint64_t message_id_;
 };
 
 template <typename T>
